refactor: year, month object instead of Date (#31)
